simpler approach keeping existing mod.rs
