+++ mqtt/mqttd/src/broker/bootstrap/mod.rs
@@ -2,7 +2,7 @@
 mod edgehub;
 
 #[cfg(feature = "edgehub")]
-pub use edgehub::{broker, config, run};
+pub use edgehub::{broker, config, start_server, start_sidecars};
 
 #[cfg(all(not(feature = "edgehub"), feature = "generic"))]
 mod generic;
